Move mesh conversion utils to `mesh_utils`
